+++ youlai-mall-service-api/yshop-pms-api/src/main/java/com/fly4j/yshop/pms/feign/app/factory/PmsAppFeignFallbackFactory.java
@@ -14,7 +14,7 @@ public class PmsAppFeignFallbackFactory implements FallbackFactory<PmsAppFeign> 
     public PmsAppFeign create(Throwable throwable) {
         return new PmsAppFeign() {
             @Override
-            public R getSkuById(Long id) {
+            public R getSkuById(Integer id) {
                 return null;
             }
